+++ CHANGES.rst
@@ -6,7 +6,9 @@ There's a frood who really knows where his towel is.
 2.0b4 (unreleased)
 ^^^^^^^^^^^^^^^^^^
 
-- Nothing changed yet.
+- Do not try to create scales on news article with no lead image;
+  this was causing issues in some Collection view methods and in collective.cover's Collection tile (fixes `#178`_).
+  [hvelarde]
 
 
 2.0b3 (2016-09-12)
@@ -161,3 +163,4 @@ There's a frood who really knows where his towel is.
 .. _`#161`: https://github.com/collective/collective.nitf/issues/161
 .. _`#169`: https://github.com/collective/collective.nitf/issues/169
 .. _`#175`: https://github.com/collective/collective.nitf/issues/175
+.. _`#178`: https://github.com/collective/collective.nitf/issues/178